🍎 Room Dao Delete Query 반환값 수정

+++ data/src/main/java/com/example/data/repository/datasource/LocalDataSource.kt
@@ -12,5 +12,5 @@ interface LocalDataSource {
 
     suspend fun deleteAll()
 
-    suspend fun deleteMovie(movieEntity: MovieEntity): Long
+    suspend fun deleteMovie(movieEntity: MovieEntity): Int
 }
